[#74718] Upgrade logs appearance

+++ src/preprocessing/bisect.sh
@@ -3,15 +3,17 @@
 fcstd_path=$1
 designs=$2
 ITERATION=0
-# Prepare
 cp "$designs"/config.json "$designs"/temporary_config.json
-rm "$designs"/bisect_log.csv
+[ -f "$designs"/bisect_log.csv ] && rm "$designs"/bisect_log.csv
 while true; do
+
     iteration=$(("$iteration"+1))
+    echo ""
+    echo "------------------ #$iteration Iteration "
     export ITERATION="$iteration"
     tpre get-coef "$fcstd_path" "$designs"/temporary_config.json
     tpre parse-fcstd "$fcstd_path" "$designs" "$designs"
-    tpre prepare "$designs"/FEMMeshGmsh.inp True
+    tpre prepare "$designs"/FEMMeshGmsh.inp True > /dev/null 2>&1
     export OMP_NUM_THREADS=12
     ccx "$designs"/FEMMeshGmsh > /dev/null 2>&1
     ccx2paraview "$designs"/FEMMeshGmsh.frd vtk > /dev/null 2>&1
